Create album: Separate upload file
