+++ packages/graphql/schemas/album/schema.graphql
@@ -1,11 +1,3 @@
-type File {
-  id: String!
-  path: String!
-  filename: String!
-  mimetype: String!
-  encoding: String!
-}
-
 type Album {
   id: Int
   name: String
@@ -33,5 +25,4 @@ type Query {
 
 type Mutation {
   createAlbum(data: AlbumInput): Album
-  uploadFile(file: Upload!): File!
 }
