+++ packages/graphql/schemas/image/resolver.ts
@@ -1,0 +1,7 @@
+import { Query } from "./image.query";
+import { Mutation } from "./image.mutation";
+
+export const resolver = {
+  // Query: Query,
+  Mutation: Mutation,
+};
